pyproject.toml
[tool.poetry]
name = "shg_simulation"
version = "0.1.0"
description = ""
keywords = [""]
authors = [
  "Jacob Duffy <jod0007@auburn.edu>",
  "Hussam Mustafa <hnm0037@auburn.edu>",
  "Brian Opatosky <bbo0007@auburn.edu>",
  "Chunli Tang <chunli.tang@auburn.edu>",
]
readme = "README.md"
license = "LICENSE.txt"

[tool.poetry.dependencies]
python = "^3.10"
pyqt6 = "^6.7.0"
pandas = "^2.2.2"
numpy = "^2.0.0"
matplotlib = "^3.9.1"
scipy = "^1.14.0"
cffi = "^1.16.0"
markdown = "^3.6"
setuptools = "^72.1.0"
pyyaml = "^6.0.2"
requests = "^2.32.3"
gitpython = "^3.1.43"
mp-api = "^0.41.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
shg_gui = "shg_simulation.src.startup:main"
